+++ src/ts/tutorials/multirate/ex3.c
@@ -169,7 +169,7 @@ int main(int argc,char **argv)
   /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
      Print norm2 error
      - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
-  PetscCall(PetscPrintf(PETSC_COMM_WORLD,"l2 error norm: %g\n", error));
+  PetscCall(PetscPrintf(PETSC_COMM_WORLD,"l2 error norm: %g\n", (double)error));
 
   /* - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
      Free work space.  All PETSc objects should be destroyed when they are no longer needed.
